First functional commit of iterqueue
This is the first commit with actual content, and should be in working
order for a first version. I am marking this as version 0.0.1 as it
has not been used in any project yet, so interfaces may change later.

Unit testing does show this as working, but please use caution.